+++ hermitian-XRCC/mains/xr1-ccsd.py
@@ -40,22 +40,29 @@ import combo_diagram
 from   get_ints import get_ints
 from precontract import precontract
 
-# needed for unpickling?!
-class empty(object):  pass     # Basically just a dictionary class
+class empty(object):  pass     # needed for unpickling - remove when all Be-states drivers updated to use struct instead
 
 #torch.set_num_threads(4)
 #tensorly.set_backend("pytorch")
+
+global_timings = timer()
+matrix_timings = timer()
+integral_timings = timer()
+diagram_timings = timer()
+precontract_timings = timer()
+qode.math.tensornet.initialize_timer()
+qode.math.tensornet.tensorly_backend.initialize_timer()
+
+global_timings.start()
 
 #########
 # Load data
 #########
 
-timings = timer()    # starts the overall clock
-
 # Information about the Be2 supersystem
 n_frag       = 2
 displacement = float(sys.argv[1])
-states       = "atomic_states/rho/{}.pkl".format(sys.argv[2])
+states       = "rho/{}.pkl".format(sys.argv[2])
 project_core = True
 if len(sys.argv)==4:
     if sys.argv[3]=="no-proj":
@@ -69,23 +76,23 @@ for m in range(int(n_frag)):
     for elem,coords in Be.atoms:  coords[2] += m * displacement    # displace along z
     BeN += [Be]
 print("get_ints ...")
-symm_ints, bior_ints, nuc_rep = get_ints(BeN, project_core)
+symm_ints, bior_ints, nuc_rep = get_ints(BeN, project_core, integral_timings)
 print("done")
 
 # The engines that build the terms
 BeN_rho = [frag.rho for frag in BeN]   # diagrammatic_expansion.blocks should take BeN directly? (n_states and n_elec one level higher)
-contract_cache = precontract(BeN_rho, symm_ints.S, timings)
-S_blocks       = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=symm_ints.S,                               diagrams=S_diagrams,  contract_cache=contract_cache, timings=timings)
-St_blocks_symm = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, T=symm_ints.T),      diagrams=St_diagrams, contract_cache=contract_cache, timings=timings)
-Su_blocks_symm = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, U=symm_ints.U),      diagrams=Su_diagrams, contract_cache=contract_cache, timings=timings)
-Sv_blocks_symm = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=symm_ints.V),      diagrams=Sv_diagrams, contract_cache=contract_cache, timings=timings)
-St_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, T=bior_ints.T),      diagrams=St_diagrams, contract_cache=contract_cache, timings=timings)
-Su_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, U=bior_ints.U),      diagrams=Su_diagrams, contract_cache=contract_cache, timings=timings)
-Sv_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=bior_ints.V),      diagrams=Sv_diagrams, contract_cache=contract_cache, timings=timings)
-Sv_blocks_half = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=bior_ints.V_half), diagrams=Sv_diagrams, contract_cache=contract_cache, timings=timings)
-Sv_blocks_diff = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=bior_ints.V_diff), diagrams=Sv_diagrams, contract_cache=contract_cache, timings=timings)
+contract_cache = precontract(BeN_rho, symm_ints.S, precontract_timings)
+S_blocks       = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=symm_ints.S,                               diagrams=S_diagrams,  contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+St_blocks_symm = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, T=symm_ints.T),      diagrams=St_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+Su_blocks_symm = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, U=symm_ints.U),      diagrams=Su_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+Sv_blocks_symm = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=symm_ints.V),      diagrams=Sv_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+St_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, T=bior_ints.T),      diagrams=St_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+Su_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, U=bior_ints.U),      diagrams=Su_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+Sv_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=bior_ints.V),      diagrams=Sv_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+Sv_blocks_half = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=bior_ints.V_half), diagrams=Sv_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
+Sv_blocks_diff = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, V=bior_ints.V_diff), diagrams=Sv_diagrams, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
 
-Combo_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, T=bior_ints.T, U=bior_ints.U, V=bior_ints.V),      diagrams=combo_diagram, contract_cache=contract_cache, timings=timings)
+Combo_blocks_bior = diagrammatic_expansion.blocks(densities=BeN_rho, integrals=struct(S=symm_ints.S, T=bior_ints.T, U=bior_ints.U, V=bior_ints.V),      diagrams=combo_diagram, contract_cache=contract_cache, timings=diagram_timings, precon_timings=precontract_timings)
 
 
 # charges under consideration
@@ -99,6 +106,9 @@ dimer_charges = {
                 }
 all_dimer_charges = [(0,0), (0,+1), (0,-1), (+1,0), (+1,+1), (+1,-1), (-1,0), (-1,+1), (-1,-1)]
 
+global_timings.record("setup")
+global_timings.start()
+
 #########
 # Build and test
 #########
@@ -111,18 +121,18 @@ for m in [0,1]:
                           1: [
                               "t00"
                              ]
-                         }, m, monomer_charges)
+                         }, m, monomer_charges, matrix_timings)
     H1_m += XR_term.monomer_matrix(Su_blocks_symm, {
                           1: [
                               "u000"
                              ]
-                         }, m, monomer_charges)
+                         }, m, monomer_charges, matrix_timings)
 
     H1_m += XR_term.monomer_matrix(Sv_blocks_symm, {
                           1: [
                               "v0000"
                              ]
-                         }, m, monomer_charges)
+                         }, m, monomer_charges, matrix_timings)
     H1 += [H1_m]
 
 
@@ -136,7 +146,7 @@ S2     = XR_term.dimer_matrix(S_blocks, {
                         2: [
                             "s01"
                            ]
-                       },  (0,1), all_dimer_charges)
+                       },  (0,1), all_dimer_charges, matrix_timings)
 
 S2inv = qode.math.precise_numpy_inverse(S2)
 
@@ -151,7 +161,7 @@ S2H2  = XR_term.dimer_matrix(St_blocks_symm, {
                        2: [
                            "t01"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 S2H2 += XR_term.dimer_matrix(Su_blocks_symm, {
                        1: [
                            "u000"
@@ -160,7 +170,7 @@ S2H2 += XR_term.dimer_matrix(Su_blocks_symm, {
                            "u100",
                            "u001", "u101"
                           ]
-                     }, (0,1), all_dimer_charges)
+                     }, (0,1), all_dimer_charges, matrix_timings)
 
 S2H2 += XR_term.dimer_matrix(St_blocks_bior, {
                        2: [
@@ -168,7 +178,7 @@ S2H2 += XR_term.dimer_matrix(St_blocks_bior, {
                            #->#"s01t10", 
                            "s01t01"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 S2H2 += XR_term.dimer_matrix(Su_blocks_bior, {
                        2: [
                            "s01u000", "s01u011",
@@ -176,7 +186,7 @@ S2H2 += XR_term.dimer_matrix(Su_blocks_bior, {
                            #->#"s01u010", "s01u110", 
                            "s01u001", "s01u101"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 
 print("build S2H2 (2e)")
 
@@ -187,7 +197,7 @@ print("build S2H2 (2e)")
 #                       2: [
 #                           "v0101", "v0010", "v0111", "v0011"
 #                          ]
-#                      }, (0,1), all_dimer_charges)
+#                      }, (0,1), all_dimer_charges, matrix_timings)
 
 S2H2 += XR_term.dimer_matrix(Sv_blocks_diff, {
                        1: [
@@ -196,7 +206,7 @@ S2H2 += XR_term.dimer_matrix(Sv_blocks_diff, {
                        2: [
                            "v0101", "v0010", "v0111", "v0011"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 
 S2H2 += XR_term.dimer_matrix(Sv_blocks_bior, {
                        2: [
@@ -205,18 +215,16 @@ S2H2 += XR_term.dimer_matrix(Sv_blocks_bior, {
                            #->#"s01v1000", 
                            "s01v1100", "s01v0010", "s01v0111"#, "s01v0011"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 
 
 S2H2 += XR_term.dimer_matrix(Combo_blocks_bior, {
                        2: [
                           "combo"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 
 
-
-print("Apply H")
 
 H2blocked = S2inv @ S2H2
 
@@ -224,20 +232,25 @@ H2blocked -= XR_term.dimer_matrix(St_blocks_symm, {
                        1: [
                            "t00"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
 H2blocked -= XR_term.dimer_matrix(Su_blocks_symm, {
                        1: [
                            "u000"
                           ]
-                     }, (0,1), all_dimer_charges)
+                     }, (0,1), all_dimer_charges, matrix_timings)
 
 H2blocked -= XR_term.dimer_matrix(Sv_blocks_symm, {
                        1: [
                            "v0000"
                           ]
-                      }, (0,1), all_dimer_charges)
+                      }, (0,1), all_dimer_charges, matrix_timings)
+
+global_timings.record("build")
+global_timings.start()
 
 
+
+print("Apply H")
 
 # well, this sucks.  reorder the states
 dims0 = [BeN[0].rho['n_states'][chg] for chg in [0,+1,-1]]
@@ -268,4 +281,14 @@ E, T = excitonic.ccsd((H1,[[None,H2],[None,None]]), out, resources)
 E += sum(nuc_rep[m1,m2] for m1 in range(n_frag) for m2 in range(m1+1))
 out.log("\nTotal Excitonic CCSD Energy (test) = ", E)
 
-timings.print()
+global_timings.record("apply")
+
+
+
+global_timings.print("GLOBAL")
+matrix_timings.print("MATRIX")
+integral_timings.print("INTEGRALS")
+diagram_timings.print("DIAGRAMS")
+precontract_timings.print("PRECONTRACTIONS")
+qode.math.tensornet.print_timings("TENSORNET COMPONENTS")
+qode.math.tensornet.tensorly_backend.print_timings("TENSORLY BACKEND")
